CORE-1450 Fix Step.Config.Parameters sort order
Use a stable sort for sorting params,
which fixes the Arguments order used for the command-line.
Particularly important for multi-input params,
which can all have the same param order,
but only the 1st will have an argument option ("name")
which can't be shuffled into the middle of the list of inputs.